Revert "Merge branch 'main' of https://github.com/grafana/grafana into feature/Russian"
This reverts commit bb37ce3fe0c47be1b34d3851dc847ce98a92b872, reversing
changes made to bdf0976781d72b153bce81de42354c8524949208.

+++ pkg/components/dashdiffs/compare.go
@@ -34,7 +34,7 @@ type Options struct {
 
 type DiffTarget struct {
 	DashboardId      int64
-	Version          int64
+	Version          int
 	UnsavedDashboard *simplejson.Json
 }
